Towards #812: carve out rank_based_allocate() to its own module
