+++ df_py/volume/reward_calculator.py
@@ -1,20 +1,18 @@
 from datetime import datetime
-from typing import Dict, List, Tuple, Union
+from typing import Dict, List, Tuple
 
 import numpy as np
-import scipy
 from enforce_typing import enforce_types
 
 from df_py.predictoor.queries import query_predictoor_contracts
 from df_py.util.constants import (
     DEPLOYER_ADDRS,
-    MAX_N_RANK_ASSETS,
-    PREDICTOOR_MULTIPLIER,
     RANK_SCALE_OP,
     TARGET_WPY,
 )
 from df_py.volume import cleancase as cc
 from df_py.volume import to_usd
+from df_py.volume.rank import rank_based_allocate
 
 
 def freeze_attributes(func):
@@ -178,7 +176,7 @@ class RewardCalculator:
                     S[self.C[j], j] *= 2.0
         # perc_per_j
         if self.do_rank:
-            perc_per_j = self._rank_based_allocate()
+            perc_per_j = rank_based_allocate(self.V_USD)
         else:
             perc_per_j = self.V_USD / np.sum(self.V_USD)
 
@@ -227,67 +225,6 @@ class RewardCalculator:
         assert sum1 <= self.OCEAN_avail * (1 + tol), (sum2, self.OCEAN_avail, R)
 
         return R
-
-    @freeze_attributes
-    @enforce_types
-    def _rank_based_allocate(
-        self,
-        max_n_rank_assets: int = MAX_N_RANK_ASSETS,
-        rank_scale_op: str = RANK_SCALE_OP,
-        return_info: bool = False,
-    ) -> Union[np.ndarray, tuple]:
-        """
-        @return
-        Always return:
-          perc_per_j -- 1d array of [chain_nft j] -- percentage
-
-        Also return, if return_info == True:
-          ranks, max_N, allocs, I -- details in code itself
-        """
-        if len(self.V_USD) == 0:
-            return np.array([], dtype=float)
-        if min(self.V_USD) <= 0.0:
-            raise ValueError(
-                f"each nft needs volume > 0. min(self.V_USD)={min(self.V_USD)}"
-            )
-
-        # compute ranks. highest-DCV is rank 1. Then, rank 2. Etc
-        ranks = scipy.stats.rankdata(-1 * self.V_USD, method="min")
-
-        # compute allocs
-        N = len(ranks)
-        max_N = min(N, max_n_rank_assets)
-        allocs = np.zeros(N, dtype=float)
-        I = np.where(ranks <= max_N)[0]  # indices that we'll allocate to
-        assert len(I) > 0, "should be allocating to *something*"
-
-        if rank_scale_op == "LIN":
-            allocs[I] = max(ranks[I]) - ranks[I] + 1.0
-        elif rank_scale_op == "SQRT":
-            sqrtranks = np.sqrt(ranks)
-            allocs[I] = max(sqrtranks[I]) - sqrtranks[I] + 1.0
-        elif rank_scale_op == "POW2":
-            allocs[I] = (max(ranks[I]) - ranks[I] + 1.0) ** 2
-        elif rank_scale_op == "POW4":
-            allocs[I] = (max(ranks[I]) - ranks[I] + 1.0) ** 4
-        elif rank_scale_op == "LOG":
-            logranks = np.log10(ranks)
-            allocs[I] = max(logranks[I]) - logranks[I] + np.log10(1.5)
-        else:
-            raise ValueError(rank_scale_op)
-
-        # normalize
-        perc_per_j = allocs / sum(allocs)
-
-        # postconditions
-        tol = 1e-8
-        assert (1.0 - tol) <= sum(perc_per_j) <= (1.0 + tol)
-
-        # return
-        if return_info:
-            return perc_per_j, ranks, max_N, allocs, I
-
-        return perc_per_j
 
     @freeze_attributes
     @enforce_types
